move longpress functionality to extension class, update images, use provided menu items where possible, drop menu items we don't want
