+++ DuckDuckGo/en.lproj/Localizable.strings
@@ -36,7 +36,7 @@
 "action.title.bookmark" = "Bookmark";
 "action.title.newTab" = "New Tab";
 "action.title.newTabForUrl" = "Open in New Tab";
-"action.title.newBackgroundTabForUrl" = "Open in Background Tab";
+"action.title.newBackgroundTabForUrl" = "Open in Background";
 "action.title.forgetAll" = "Clear Tabs and Data";
 "action.title.forgetAllDone" = "Tabs and data cleared";
 "action.title.open" = "Open";
